update package-lock after build
We need to commit changes to package.json and package-lock.json back to
the repository. Since we install dependencies with `npm ci`, the only
update to package-lock should be updating the version of this package.
If we we don't commit that, the next dev to run `npm install` in the
repo will have a package-lock.json change that updates the version to...
the N-1 version.